==> 2_descriptives.Rmd <==
---
title: "Descriptive Analyses for Air Pollution & Dementia"
author: "Magali Blanco"
date: ' `r Sys.time()` '
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

# Data

```{r, setup, include=F}
# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))
}

knitr::opts_chunk$set(echo = F, cache=F, cache.comments = F, message = F, warning = F, 
                      tidy.opts=list(width.cutoff=60), tidy=TRUE   
                      )  

pacman::p_load(kableExtra, tidyverse, table1, lubridate,
               ggpubr, #stat_cor() for ggplot
               GGally, # ggpairs()
               ggmap, sf, ggspatial, cowplot, #mapping...adding scales, N arrows
               PRISMAstatement, # sample size left
               survival, survminer
               )    
 
# ggplot settings
theme_set(theme_bw())
theme_update(legend.position = "bottom")

set.seed(1)

image_path <- file.path("..", "manuscript", "images")
 
# Load files 
source("functions.R")
sur_person <- readRDS(file.path("Data", "Output", "sur_person.rda"))

drop_pollutants <- c("uv_bc", "blue_bc", "green_bc", "red_bc",
                     paste0("a", 1:3), paste0("m_", c("a", "b", "c", "d", "e", "f")),
                     "nox", "co2"
                     )

# want to drop last 2 years w/ administrative censoring - incidence is artificially high b/c person w/o visits for 
start_of_bad_yrs <- readRDS(file.path("Data", "Output", "1_prep_data", "start_of_bad_yr.rda"))

sur <- readRDS(file.path("Data", "Output", "sur.rda")) %>% 
  filter(!pollutant %in% drop_pollutants,
         exposure_year < start_of_bad_yrs,
         )

exclusion_table <- readRDS(file.path("Data", "Output", "exclusion_table.rda")) 


```

```{r}

# https://cran.r-project.org/web/packages/PRISMAstatement/vignettes/exclusionflowcharts.html 


#  last row looks weird
# persons
flow_exclusions(
  incl_counts = exclusion_table$persons,
  total_label = "Full Cohort",
  incl_labels = exclusion_table$description[2:length(exclusion_table$description)],
  percent_of_total = T,
  format_args = list(big.mark  = ",")
  )

# person-years
flow_exclusions(
  incl_counts = exclusion_table$person_years,
  total_label = "Full Cohort",
  incl_labels = exclusion_table$description[2:length(exclusion_table$description)],
  percent_of_total = T,
  format_args = list(big.mark  = ",")
  )

```


```{r}
exclusion_table2 <- exclusion_table %>%
  mutate(persons_pct = round(persons/first(persons)*100),
         person_years_pct = round(person_years/first(person_years)*100),
         
         persons_pct_dropped = round(persons_dropped/first(persons)*100),
         person_years_pct_dropped = round(person_years_dropped/first(person_years)*100),
         
         persons = paste0(persons, " (", persons_pct, "%)"),
         persons_dropped = paste0(persons_dropped, " (", persons_pct_dropped, "%)"),
         
         person_years = paste0(person_years, " (", person_years_pct, "%)"),
         person_years_dropped = paste0(person_years_dropped, " (", person_years_pct_dropped, "%)"),
         ) %>%
  select(description, persons, persons_dropped, person_years, person_years_dropped)

```

# --> why does T1 have time nera 0 but start/end visits say ~2 yrs differnce? 
1. don't trust intageage & last_visit_age variables
2. person-years may have been dropped for some people b/c of low exposure coverage. when. look in 1_prep_data.R (~line 243), there are only 3 ppl with only 1 row (exposure year) of data. Thus, some time must be dropped later for people.

```{r}

# sur %>% 
#   filter(pollutant=="ufp_10_42", exposure_duration==10, model=="MM") %>%
#   group_by(study_id) %>%
#   mutate(n=n()) %>% 
#   filter(n<=1)  


```


# Table 1
 
```{r}
#calculate whether individuals were above/below 10-year avg MM pollutant concentration at baseline
baseline_conc_avg <- sur %>%
  filter(exposure_duration==10,
         model== "MM",
         pollutant %in% c("ufp_10_42", "bc", "no2")) %>% 
  group_by(study_id) %>%
  filter(exposure_year==min(exposure_year)) %>% 
  group_by(pollutant, exposure_year) %>%
  mutate(
    mean_baseline_conc = mean(pollutant_prediction),
    above_average = pollutant_prediction >= mean_baseline_conc) %>%
  ungroup() %>% 
  select(study_id, age_start_exposure, exposure_year, pollutant, pollutant_prediction, above_average) %>%
  pivot_wider(names_from = pollutant, values_from = c(pollutant_prediction, above_average))

sur_person_t1 <- left_join(sur_person, baseline_conc_avg) %>%
  mutate(apoe_available = !is.na(apoe))

 
# Table1 tutorials: 
## https://cran.r-project.org/web/packages/table1/table1.pdf 
## https://cran.r-project.org/web/packages/table1/vignettes/table1-examples.html

# # round statistics
# my.render.cont <- function(x) {
#     with(stats.apply.rounding(stats.default(x), digits=2), c("",
#         "Mean (SD)"=sprintf("%s (%s)", MEAN, SD)))
# }
# my.render.cat <- function(x) {
#     c("", sapply(stats.default(x), function(y) with(y,
#         sprintf("%d (%0.0f%%)", FREQ, PCT))))
# }

  
t1_labels <- list(
  labels = list(
    male = "Sex",
    #intakeage = "Age at Intake",
    age_start_exposure = "Age at Baseline",
    follow_up_years = "Follow-Up Duration",
    corrected_anydementia = "All-Cause Dementia Cases",
    ad_nincds = "Alzheimer's Disease Dementia Cases",
    vad_dementia = "Vascular Dementia",
    mixed_dementia = "Mixed Dementia",
    other_dementia = "Other Dementia",
    non_ad_dementia = "Non-Alzheimer's Disease Dementia Cases",
    #birth_cohort = "Birth Cohort",
    race_white = "Race",
    degree = "Degree",
    nses_z_cx = "Neighborhood Disadvantage Index (NDI)",
    apoe = "APOE Genotype",
    apoe_available = "Included in Main Analysis",
    model_wt = "IPW for Modeling Cohort",
    above_avg_ufp_10_42 = "Above Avg (PNC)",
    pollutant_prediction_ufp_10_42 = "PNC (pt/cm3) Exposure",
    pollutant_prediction_no2 = "NO2 (ppb) Exposure",
    pollutant_prediction_bc = "BC (ng/m3) Exposure" #,  pollutant_prediction_pm25 = "PM2.5 (ug/m3)"
    ),
  units = list(
    #intakeage = "Years"
    age_start_exposure = "Years"),
  categoricals = list(
    male = list('1' = "Male",
                '0' = "Female"),
    corrected_anydementia = list('1' = "Yes",
                                 '0' = "No"),
    vad_dementia = list('1' = "Yes", 
                        '0' = "No"),
    mixed_dementia = list('1' = "Yes", 
                        '0' = "No"),
    other_dementia = list('1' = "Yes", 
                        '0' = "No"),
    non_ad_dementia = list('1' = "Yes", 
                        '0' = "No"),
    ad_nincds = list('1' = "Yes",
                     '0' = "No"
                                 ),
    race_white = list('1' = "White",
                      '0' = "People of Color"),
    degree = list('0' = "None",
                  '1' = "GED/High School",
                  '3' = "Bachelor's",
                  '4' = "Master's",
                  '5' = "Doctorate",
                  '6' = "Other"),
    apoe = list('1' = "1+ e4 Allele",
                '0' = "No e4 Allele"),
    apoe_available = list('TRUE' = "Included in Main Analysis",
                          'FALSE' = "Not Included in Main Analysis"),
    above_average_ufp_10_42 = list('FALSE' = "PNC Below Avg",
                                   'TRUE' = "PNC Above Avg")
    )#, groups=list("", "PNC Conc", "")
  )

# saveRDS(t1_labels, file.path("Data", "Output", "table1_labels.rda"))

print("Baseline cohort characteristics stratified by whether 10-year average PNC concentration at baseline was below or above that year's average.")
# Main Table 1
t1read(filter(sur_person_t1, apoe_available==TRUE), t1_labels) %>%
  table1(~age_start_exposure + follow_up_years + male + race_white +degree + nses_z_cx + #income_cat + 
           apoe + model_wt + corrected_anydementia + ad_nincds + non_ad_dementia, #+ pollutant_prediction_ufp_10_42 + pollutant_prediction_no2 + pollutant_prediction_bc, #|above_average_ufp_10_42, 
         data=.)

```


# Prediction Maps


```{r, eval=F}
# don't need these anymore??

## mapping
lat_long_crs <- 4326  
m_crs <- 32148 #meters

monitoring_area <- readRDS(file.path("~", "OneDrive - UW", "Documents", "Post Doc", "Study Projects", "ACT TRAP MM", "1. Our Campaign",  "Our Campaign R", "Data", "Output", "GIS", "monitoring_area_shp.rda")) 
st_area <- read_sf(file.path("~", "OneDrive - UW", "Documents", "School", "PhD", "Dissertation", "GIS", "Shapefiles", "Study area", "Spatiotemporal", "st_area.shp")) #%>% st_transform(project_crs)

#make box little bigger than monitoring area
bbox <- st_bbox(st_transform(st_buffer(st_transform(st_area, m_crs), 40e3), lat_long_crs))
names(bbox) <- c("left", "bottom", "right", "top")
map0 <- suppressMessages(get_stamenmap(bbox = bbox, zoom = 11, #14, #11, 
                                       maptype = "toner-lite" #"terrain" #
                                       )) %>%
  ggmap(ggmap = ., darken = c(0.5, "white")) + theme_void()

 
main_map <- map0 + 
  geom_sf(data=st_area, inherit.aes = F, aes(fill = "Spatiotemporal"), 
          alpha=0.1, lwd = 0.1,) + 
  geom_sf(data=monitoring_area, inherit.aes = F, aes(fill = "Mobile Monitoring"), 
          alpha=0.1, lwd = 0.1) +
  
  annotation_scale(location = "tl") +
  annotation_scale(location = "tl", unit_category ="imperial", pad_y = unit(0.55, "cm")) +
  annotation_north_arrow(location = "tl", which_north = "true", pad_y = unit(0.5, "in"),
                           style = north_arrow_fancy_orienteering) +
  theme_bw() +
  theme(
    legend.justification=c(1,1),  
    legend.position=c(1,1),  
    legend.background =  element_blank()
    ) +
    coord_sf(expand = F) +
  #add attribution/reference to bottom left
    geom_text(aes(x=-Inf, y=-Inf, hjust=-0.01, vjust=-0.3, 
                  label= "Map tiles by Stamen Design, under CC BY 3.0. \nData by OpenStreetMap, under ODbL."), size=2.5) +
  labs(x = "Longitude",
       y = "Latitude",
       fill = "Prediction Area"
       ) 

# inste map 
# example code: https://geocompr.github.io/post/2019/ggplot2-inset-maps/ 

data("us_states", package = "spData")

wa_map <- us_states %>%
  filter(NAME == "Washington") %>%
  st_transform(lat_long_crs)

wa_centroid <- st_coordinates(st_centroid(wa_map))

inset_map <- ggplot() + 
  geom_sf(data=wa_map, fill = "white", alpha=0.5) + 
  geom_sf(data=st_area, aes(fill = "Spatiotemporal"), lwd = 0.1, alpha = 0.2, show.legend = F) +
  geom_sf(data=monitoring_area, aes(fill = "Mobile Monitoring"), lwd = 0.1, alpha = 0.2, show.legend = F) +
  theme_void() + 
  #add "WA" label
  geom_text(aes(x = wa_centroid[1], y = wa_centroid[2]), label = "WA", size=4)
  
# combine using cowplot
ggdraw() +
  draw_plot(main_map) +
  draw_plot(inset_map, 
            # The dist along a (0,1) x- or y- axis to draw the left/bottom edge of the plot
            x = 0.65, y = 0.005, 
            # The width and height of the plot as proportion of the entire ggdraw object
            width = 0.23, height = 0.23)

ggsave(file.path(image_path, "SI", "monitoring_map.jpg"), width = 6, height = 8)

```

 
# Exposure Quality 

* **results are for 10 year exposure window from the MM model for the first pollutant listed**

* exp_wks_coverage, imp_quality, exact_coverage, and imputed_coverage

**exp_wks_coverage**.  
* Proportion of weeks of desired pollutant exposure with nonmissing location/pollutant info
* for MM and ST regions.
    * MM is slightly smaller for PM2.5 and NO2, as we would expect 
* the imputation quality and exact coverage variables are related to ALL of the known address histories, an Not necessarily the address histories that get used in the exposure models. Meaning that they are more useful/better indicators of the quality of the predictions when exposure coverage is high (e.g., 1.0), and less helpful if exposure coverage is low
* Amanda: "if coverage is low I would expect those rows to be excluded anyway.  If the coverage is low then your exposure for that row will have a really different error structure from exposure windows where coverage is high. In general, for most analyses in e.g. MESA, we exclude any rows where coverage is < 95%." 

**imp_qual10_yr_**.  
* Maximum imputation quality of the locations that fall in the relevant period of interest  
* people without valid addresses in the ST modeing region will have NAs     


```{r, results = "asis"}
indicator_variables <- c("exp_wks_coverage10_yr_MM",
                         "exact_coverage10_yr_MM",
                         "imputed_coverage10_yr_MM"#, "imp_qual10_yr_MM"
                         )
lapply(indicator_variables, function(x) {
  t <- filter(sur, 
       pollutant == "ufp_10_42", #first(pollutant),
       model == "MM",
       exposure_duration ==10
       ) %>%   
  summary_table(var = x) %>%
  kable(caption = paste("distribution of", x), digits = 3) %>%
  kable_styling()
  
  print(t)
  cat("\n")
})


filter(sur, 
       pollutant == "ufp_10_42", #first(pollutant),
       model == "MM",
       exposure_duration ==10
       ) %>%
  summarize(
    n = n(),
    imp0 = mean(imp_qual10_yr_MM==0),
    imp1 = mean(imp_qual10_yr_MM==1),
    imp2 = mean(imp_qual10_yr_MM==2),
    ) %>%  
  kable(caption = "proportion of each imp_qual10_yr_MM value (0-2)", digits = 3) %>%
  kable_styling()
  
  

```



# IPW

* distribution of weights in sur_person

* could stratify by covariates

```{r}
sur_person %>%
  drop_na(model_wt) %>%
summary_table(., "model_wt") %>% 
  kable(caption = "distribution of IPW", digits = 2) %>%
  kable_styling()

print("persons with modelw weights > 2")
table(sur_person$model_wt>2)

```


# Exposure 

## Exposure Fig/Table

figure 

```{r}

exposure_df <- sur %>%
  filter(exposure_duration==10,
         #!(grepl("PM25", pollutant, ignore.case = T) & grepl("MM|ST", model))
         !(grepl("PM25", pollutant, ignore.case = T) & grepl("MM", model))
         ) %>%
  mutate(
    model = ifelse(model== "SP", "2019 Spatiotemporal", 
                   ifelse(model=="ST", "Spatiotemporal", 
                          ifelse(model=="MM", "Mobile Monitoring",  NA))),
    model = factor(model, levels=c("Mobile Monitoring", "2019 Spatiotemporal", "Spatiotemporal"))) %>%
  label_pollutants2() %>%
  
  mutate(
    pt_range = ifelse(grepl("Particle Size", total_pnc), "Mean Size", as.character(pt_range)),
    pt_range = factor(pt_range, levels=c("10-420 nm", "10-420 nm\n(Primary Analysis)",
                                         "20-1,000 nm",
                                         "20-1,000 nm, Onroad",
                                         "10-700 nm",
                                         "10-100 nm",
                                         "10-13 nm",
                                         "13-18 nm",
                                         "18-24 nm",
                                         "24-32 nm",
                                         "32-42 nm",
                                         "42-56 nm",
                                         "56-75 nm",
                                         "75-100 nm",
                                         "100-133 nm",
                                         "133-178 nm",
                                         
                                         "20-36 nm", 
                                         "36-1,000 nm", " ",
                                         "Mean Size")))
  

#UFP
ufp_p <- exposure_df%>% 
  filter(pollutant=="UFP") %>%  
  
  ggplot(aes(x=pollutant_prediction, col=instrument, y= pt_range)) + 
  facet_wrap(~pollutant+total_pnc, scales="free", ) + 
  geom_boxplot() + 
  scale_x_log10() +
  labs(col="Instrument", x = "10yr Exposure", y="Mobile Monitoring Model") 

#non-UFP
other_p <- exposure_df%>% 
  filter(pollutant!="UFP") %>%
  
  ggplot(aes(x=pollutant_prediction, y=model)) + 
  facet_wrap(~pollutant, scales="free_x") + 
  geom_boxplot() + 
  scale_x_log10() +
  labs(col="Model", x = "10yr Exposure", y="Model")  

ggarrange(ufp_p, other_p, nrow = 2, labels = "auto", heights = c(2,1))

ggsave(file.path(image_path, "SI", "exposure_plots.png"), height = 8, width = 8)

```

table


```{r}
# sur %>%
#   filter(exp_avg0 == "exp_avg10_yr_MM",
#          pollutant %in% c("ufp_10_42", "bc", "no2")
#          ) %>%  
exposure_df %>%
  filter(grepl("NO2|BC", pollutant) | grepl("Primary Analysis", pt_range),
         model == "Mobile Monitoring"
         ) %>%

  group_by(pollutant, #Model=model
           ) %>% 

  #summary_table("pollutant_prediction") %>%
  summarize(
    py = sum(age_end_exposure - age_start_exposure), #note, rounding this changes py to 37,093, so don't do this until the end!
    df_rows = length(pollutant_prediction),
    Min = min(pollutant_prediction),
    Q25 = quantile(pollutant_prediction, 0.25),
    Mean = mean(pollutant_prediction),
    Median = median(pollutant_prediction),
    Q75 = quantile(pollutant_prediction, 0.75),
    #SD = sd(pollutant_prediction),
    #IQR = IQR(pollutant_prediction),
    Max = max(pollutant_prediction)
      ) %>%
  kable(caption = "Distribution of 10-year average MM pollutant predictions for the cohort (N=____ person-years).", 
        digits = 1, 
        #format = list(big.mark  = ",")
        ) %>%
  kable_styling()


```


```{r, eval=F}
## PM2.5 & NO2 scatterplot

print("Comparison of 10 year exposure predictions from MM and ST models")

sur %>%
  filter(exposure_duration==10) %>%
  label_pollutants() %>%
  select(study_id, exposure_year,  model, pollutant,  pollutant_prediction) %>% 
  pivot_wider(names_from = model, values_from = pollutant_prediction) %>%
  pivot_longer(cols = c(SP, ST),  names_to = "ST_model", values_to="ST_value") %>%
  drop_na() %>%
  
  ggplot(aes(x=MM, y=ST_value, col=ST_model)) + 
  facet_wrap_equal(ST_model~pollutant, scales="free") +
  geom_point(alpha=0.01, shape=21) +
  geom_smooth() + 
  geom_abline(slope = 1, intercept = 0, alpha=0.5, linetype=2) +
  theme(aspect.ratio=1) +
  #add pearson correlation
  # --> doesn't work to only show r 
  stat_cor(aes(#label = ..r.label..
    )) + 

  labs(
    x = "Mobile Monitoring Prediction",
    y = "Spatio-Temporal Prediction"
  )

#ggsave(file.path(image_path, "SI", "no2_pm25_comparisons.png"), height = 8, width = 8)


```




## Event ties

There were 85 dementia ties for the following models:  

* the specific exposures used to calculate this doesn't matter b/c ties happen on the age-scale regardless of the exposure assigned

```{r}
sur %>%
  filter(!is.na(apoe),
         model==first(model),
         exposure_duration == first(exposure_duration),
         pollutant==first(pollutant),
         dementia_now==1,
         # # only keep people in main analysis
         # exposure_year < start_of_bad_yrs,
         ) %>%  
  group_by(age_end_exposure) %>%
  summarize(N = n()) %>%
  filter(N >1) %>%
  distinct() %>% 
  summarize(event_ties = n()) 
  
```


# NDI vs Exposure

are we over adjusting for SES, especially since NDI is continuous? In some instances, are the same/few people being compared over time? for MM, this would mean small exposure changes

```{r}
sur %>%
  filter(#exposure_year < start_of_bad_yrs,
         pollutant == "ufp_10_42",
         exp_avg0 == "exp_avg10_yr_MM" #10 yr from MM
         )  %>% 
  
  ggplot(aes(x=nses_z_cx, y=pollutant_prediction, 
             #col=cal_2yr
             )) + 
  #facet_wrap(~cal_2yr) +
  geom_point(alpha=0.05) + 
  geom_smooth() + 
  geom_smooth(method = "lm", linetype=2) + 
  
  labs(title = "NDI vs 10 yr MM UFP (10-420 nm PNC) exposure prediction.\nEach dot is a person-year")
  
```

# Proportional Hazards assumption     

* apoe
* sex 
* race (White vs. People of Color)
* ten-year average PM2.5
* calendar year (two-year categories)
* education
* NDI 


Plots checking to see if the proportional hazards assumption is held.   
1. we should be able to multiply the baseline cumulative hazard function by a constant to get all of th other curves    
2. **Crossing survival curves indicate non-proportinality** though they cannot tell you (if no crossing is occurring) if the proportinal hazards assumption *is* met.

* curves that end w/ horizontal step indicate censoring events, while those that end in vertical lines indicate an event

in ggsurvplot() plots, ticks on the curves are individuals (person-years) that are censored. are horizontal ticks when there was an event And somebody else was censored/left the study?


# --> how to look at time-varying UFP/PM2.5? baseline?

#--> is cum hazard -log(esitmate)?? or -log10(estimate)? or something else?

```{r, fig.width=8, fig.height=8}
adjusted <- c( "apoe", "male", "race_white",  "degree", "nses_z_cx", "cal_2yr"#, "pm25_bin"
              )

sur_hazard <- sur %>%
  select(study_id, age_start_exposure, age_end_exposure, dementia_now, all_of(adjusted)) %>%
  distinct() %>%
  mutate(
    nses_z_cx = cut(nses_z_cx, breaks = 4, include.lowest = T),
    #cal_2yr = cut(as.numeric(as.character(cal_2yr)), breaks = 4, include.lowest = T),
    #positive_time = age_end_exposure > age_start_exposure
  )  

#create a survival object
s.dem <- Surv(
  time = sur_hazard$age_start_exposure, 
  time2 = sur_hazard$age_end_exposure, 
  event = sur_hazard$dementia_now)


pdf(file = file.path(image_path, "Other", "cumulative_hazard_functions.pdf"))

# x=adjusted[1]
lapply(adjusted, function(x) {

  print(x)
  fit <- survfit(as.formula(paste("s.dem ~", x)), data = sur_hazard) %>%
    broom::tidy() %>%
    # --> correct???
    mutate(cum_haz = -log10(estimate))
  
  p <- ggplot(fit, aes(x=time, y=cum_haz, col=strata)) + 
    geom_line() +
    #geom_ribbon(aes(ymin=conf.low, ymax=conf.high), alpha=.25)
    labs(title = x, 
         subtitle = "this should be the cumulative hazard (vs surivival)?"
         )
  
  
  # #--> has error for some reason
  # #cumulative hazards plot
  # p <- ggsurvplot(#data = sur_hazard, 
  #                 fit = fit,
  #            fun = "cumhaz", conf.int = F, 
  #            #plot faster. #don't draw ticks along curve
  #            censor=F,
  #            risk.table = T,
  #            xlim=c(65,110),
  #            title = x
  #            )

  print(p)

})

dev.off()

```
