==> 4_model_results.Rmd <==
---
title: "Hazard ratios for dementia incidence by TRAP exposure"
author: "Magali Blanco"
date: ' `r Sys.time()` '
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

# Data

```{r, setup, include=F}
# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))
}

knitr::opts_chunk$set(echo = F, cache=F, cache.comments = F, message = F, warning = F, 
                      tidy.opts=list(width.cutoff=60), tidy=TRUE)  

pacman::p_load(kableExtra, 
               tidyverse)    
 
source("functions.R")
output_data_path <- file.path("Data", "Output")

# ggplot settings
theme_set(theme_bw())
theme_update(legend.position = "bottom")

set.seed(1)

image_path <- file.path("..", "manuscript", "images")
#if(!file.exists(output_data_path)) {dir.create(output_data_path)} 
if(!file.exists(file.path(image_path, "Other", "HR Plots"))) {dir.create(file.path(image_path, "Other", "HR Plots"))} 
if(!file.exists(file.path(image_path, "Other", "PM2.5"))) {dir.create(file.path(image_path, "Other", "PM2.5"))} 

#modeled units
load(file.path(output_data_path, "trap_model_units.rda"))

# Load files 

hrs0 <- readRDS(file.path("Data", "Output", "hazard_ratios.rda")) %>%
  filter(!grepl("_bc|co2", pollutant_predictors))

main_models <- str_subset(unique(hrs0$description), "Main" )

hrs <- hrs0 %>%
  filter(grepl("MM_|SP_|ST_", covariate),
         !grepl("Interaction", description)) %>% 
  mutate( 
    description = ifelse(covariate=="SP_no2", "Spatiotemporal 2019",
                         ifelse(covariate=="ST_no2", "Spatiotemporal", description
                         )),
    
    covariate = paste(substr(covariate, 4, str_length(covariate)), substr(covariate, 1, 2), sep = "_"),
    covariate = substr(covariate, 1, str_length(covariate)-3),
    model = ifelse(str_count(pollutant_predictors, "\\+")==0, "Single Pollutant",
                   ifelse(str_count(pollutant_predictors, "\\+")==1, "Two Pollutant\n(+PM2.5)",
                          ifelse(str_count(pollutant_predictors, "\\+")==2, "Three Pollutant",
                                 ifelse(str_count(pollutant_predictors, "\\+")==3, "Four Pollutant\n(PNC+BC+NO2+PM2.5)", NA)))),
    model = factor(model, levels = c("Single Pollutant", "Two Pollutant\n(+PM2.5)", "Three Pollutant", "Four Pollutant\n(PNC+BC+NO2+PM2.5)")),
    
    ) %>%  
  rename(pollutant = covariate) %>% 
  label_pollutants() %>%
  mutate(hr_label = paste0(hr, " (", lower95, "-", upper95, ")"),
         pollutant = gsub(pattern = "PNC (", replacement = paste0("PNC (", pnc_units, " "), x = pollutant, fixed = T), 
         pollutant = gsub(pattern = "BC (", replacement = paste0("BC (", bc_units, " "), x = pollutant, fixed = T), 
         pollutant = gsub(pattern = "NO2 (", replacement = paste0("NO2 (", no2_units, " "), x = pollutant, fixed = T), 
         pollutant = gsub(pattern = "PM2.5 (", replacement = paste0("PM2.5 (", pm25_units, " "), x = pollutant, fixed = T),
         pollutant = gsub(", ", "\n", pollutant),
         pollutant = gsub("Median Pt Size", "Mean Particle nm", pollutant),
         outcome = ifelse(outcome=="dementia_now", "All-Cause Dementia", 
                          ifelse(outcome == "ad_now", "AD Dementia",
                                 ifelse(outcome == "non_ad_now", "Non-AD Dementia", NA))),
         outcome = factor(outcome, levels = c("All-Cause Dementia", "AD Dementia", "Non-AD Dementia")),
         pollutant_predictors = str_to_upper(pollutant_predictors),
         pollutant_predictors = gsub("UFP_10_42|UFP_10_70|UFP_20_1K|UFP_36_1K|_NS", "PNC", pollutant_predictors),
         pollutant_predictors = gsub("PM25", "PM2.5", pollutant_predictors), #) %>% View() #,
      
         
         significant = ifelse((lower95 <1 & upper95<1) | (lower95 >1 & upper95>1), "Significant", "Non-Significant" ),
        
         hr_label = paste0(format(round(hr, 2), nsmall=2), " (", format(round(lower95, 2), nsmall=2), "-", format(round(upper95, 2), nsmall=2), ")"),
         # why is there a space here??
         
         # --> hard coding HRS of 1, which is not ideal
         hr_label = gsub(" ", "", hr_label),
         hr_label = gsub("\\(", " \\(", hr_label),
         
         #hr_label = ifelse(significant == "Significant", paste0(hr_label, "*"), hr_label),
         
         description = relevel(factor(description), ref = main_models),
         
         exposure_duration = factor(exposure_duration, levels = rev(unique(exposure_duration))),
         exposure_model = gsub("SP", "2019 ST", exposure_model),
         exposure_model = gsub("MM", "2019 MM", exposure_model),
         )  
  
#hrs %>% View()

pollutant_levels <- c(paste0("PNC (", pnc_units, " pt/cm3)\nNanoScan"), paste0("PNC (", pnc_units, " pt/cm3) NanoScan"), #"PNC (1000 pt/cm3) NanoScan",
                      paste0("BC (", bc_units, " ng/m3)"), 
                      paste0("NO2 (", no2_units, " ppb)"), 
                      paste0("PM2.5 (", pm25_units, " ug/m3)"),
                      
                      #NS bins
                      "PNC\n10-100 nm (pt/cm3)", 
                      "PNC\n10-13 nm (pt/cm3)",
                      "PNC\n13-18 nm (pt/cm3)",
                      "PNC\n18-24 nm (pt/cm3)" ,
                      "PNC\n24-32 nm (pt/cm3)",
                      "PNC\n32-42 nm (pt/cm3)",
                      "PNC\n42-56 nm (pt/cm3)",
                      "PNC\n56-75 nm (pt/cm3)",
                      "PNC\n75-100 nm (pt/cm3)",
                      "PNC\n100-133 nm (pt/cm3)",
                      "PNC\n133-178 nm (pt/cm3)",
                      
                      #P-trak differences
                      "PNC\n20-36 nm (pt/cm3)",
                      "PNC\n36-1k nm (pt/cm3)",
                      
                      #other total UFP instruments
                      paste0("PNC (", pnc_units, " pt/cm3)\nP-TRAK"), paste0("PNC (", pnc_units, " pt/cm3)\nDiSCmini"),
                      #"PNC (1000 pt/cm3)\nP-TRAK", "PNC (1000 pt/cm3)\nDiSCmini",
                      
                      #onroad PTRAK
                      paste0("PNC (", pnc_units, " pt/cm3)\nOnroad"), #"PNC (1000 pt/cm3)\nOnroad",
                      
                      #size
                      "Median Pt Size\nDiSCmini",
                      "Mean Particle nm\nDiSCmini"
                      )
hrs <- hrs %>%
  mutate(pollutant = factor(pollutant, levels = c(pollutant_levels)))


# exploratory interaction
hrs_em <- hrs0 %>%
  filter(grepl("MM_|SP_|ST_", covariate),
         !grepl("pm25", covariate),
         grepl("Interaction", description),
         !grepl("P-Value", description),
         ) %>%
  mutate(
    significant = ifelse((lower95 <1 & upper95<1) | (lower95 >1 & upper95>1), "Significant", "Non-Significant" ),
    hr_label = paste0(round(hr,2), " (", round(lower95, 2), "-", round(upper95, 2), ")"),
    hr_label = ifelse(significant == "Significant", paste0(hr_label, "*"), hr_label)
    ) %>% 
  select(covariate, description, hr_label, hr, lower95, upper95) 

###########################################################################
# alt ST PM2.5 HRs using the same MM dataset
hrs_st_pm25 <- readRDS(file.path(output_data_path, "hazard_ratios_alternative_st_pm2.5.rda")) %>%  
  filter(grepl("pm25|no2|bc|ufp", covariate),
         # drop ST PM2.5 only model
         pollutant_predictors != "pm25"
         ) %>%
  mutate(
    covariate = paste(substr(covariate, 4, str_length(covariate)), substr(covariate, 1, 2), sep = "_"),
    covariate = substr(covariate, 1, str_length(covariate)-3),
    model = ifelse(str_count(pollutant_predictors, "\\+")==0, "Single Pollutant",
                   ifelse(str_count(pollutant_predictors, "\\+")==1, "Two Pollutant\n(+PM2.5)",
                          ifelse(str_count(pollutant_predictors, "\\+")==2, "Three Pollutant",
                                 ifelse(str_count(pollutant_predictors, "\\+")==3, "Four Pollutant\n(PNC+BC+NO2+PM2.5)", NA)))),
    model = factor(model, levels = c("Single Pollutant", "Two Pollutant\n(+PM2.5)", "Three Pollutant", "Four Pollutant\n(PNC+BC+NO2+PM2.5)")),
    ) %>%
    
    rename(pollutant = covariate) %>% 
  label_pollutants() %>%
  mutate(#hr_label = paste0(hr, " (", lower95, "-", upper95, ")"),
         pollutant = gsub(pattern = "PNC (", replacement = paste0("PNC (", pnc_units, " "), x = pollutant, fixed = T), 
         pollutant = gsub(pattern = "BC (", replacement = paste0("BC (", bc_units, " "), x = pollutant, fixed = T), 
         pollutant = gsub(pattern = "NO2 (", replacement = paste0("NO2 (", no2_units, " "), x = pollutant, fixed = T), 
         pollutant = gsub(pattern = "PM2.5 (", replacement = paste0("PM2.5 (", pm25_units, " "), x = pollutant, fixed = T),
         pollutant = gsub(", ", "\n", pollutant),
         #pollutant = gsub("Median Pt Size", "Mean Particle nm", pollutant),
         outcome = ifelse(outcome=="dementia_now", "All-Cause Dementia", 
                          ifelse(outcome == "ad_now", "AD Dementia",
                                 ifelse(outcome == "non_ad_now", "Non-AD Dementia", NA))),
         outcome = factor(outcome, levels = c("All-Cause Dementia", "AD Dementia", "Non-AD Dementia")),
         pollutant_predictors = str_to_upper(pollutant_predictors),
         pollutant_predictors = gsub("UFP_10_42|UFP_10_70|UFP_20_1K|UFP_36_1K|_NS", "PNC", pollutant_predictors),
         pollutant_predictors = gsub("PM25", "PM2.5", pollutant_predictors),
         # move PM2.5 to the right like other dataset
         
         
         significant = ifelse((lower95 <1 & upper95<1) | (lower95 >1 & upper95>1), "Significant", "Non-Significant" ),
         hr_label = paste0(format(round(hr, 2), nsmall=2), " (", format(round(lower95, 2), nsmall=2), "-", format(round(upper95,2), nsmall=2), ")"),
         exposure_duration = as.factor(exposure_duration)
         )

```


# HR Plots

* most of the significant HRs come froem the SP exposure model
* 10 yr PNC from the 2005+ exposure year models are the only significant PNC HRs   

```{r}
# dt = hrs
# description. = "Main Analysis"
# exposure_duration. = "10"

# fn returns HR plots
plot_hrs <- function(dt, description., exposure_duration., row_facet = "pollutant_predictors") {
  
  p <- dt %>%
    filter(description == description., 
           exposure_duration == exposure_duration.) %>%  
    
    ggplot(., aes(y=pollutant, col=exposure_model, shape = significant, 
                  x=hr, xmin=lower95, xmax=upper95,)) + 
    facet_grid(rows = vars(get(row_facet)), cols = vars(outcome), 
               space="free", scales="free_y", switch = "both") + 
    geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
    geom_pointrange(position = position_dodge(width = 1))+
    #geom_text(aes(label=hr_label, x=Inf, hjust=1), position = position_dodge(width = 1), size=3)+
    labs(title = paste0(description., "\n", exposure_duration., " Yr Exposure" ), 
         x= "HR", y="Pollutant Model", col = "Exposure Model", shape="")
  
  return(p)
}

```

note units:   
* PNC `r pnc_units`   
* BC `r bc_units`   
* NO2 `r no2_units`
* PM2.5 `r pm25_units`


"2005+" Models subset to individuals not yet diagnosed with dementia at the start of 2005 (vs 1994/5). Exposure models go back to 1995.


some observations   
* don't see much of an effect w/ screened p-trak and a slightly higher effect for NanoScan & DiSCmini b/c of the different particle sizes they capture? 


## Main analysis: all-cause, 10 yr exposure


```{r, eva=F}
##TEST 
# --> see if a forest plot would work easily for all 3 pollutants here?
# https://cran.r-project.org/web/packages/forestploter/vignettes/forestploter-intro.html 

pacman::p_load(grid, forestploter)

p1_t <- hrs %>%
  filter(grepl("Main", description),
        !grepl("PM2.5|DiSCmini|P-TRAK", pollutant),
        # dont include SP or ST NO2
        grepl("2019 MM", exposure_model)) %>%
  # mutate(
  #   pollutant = gsub("\nNanoScan", "", pollutant)) %>%   
  select(pollutant, hr, lower95, upper95, model, hr_label) %>% 
  mutate(
    pollutant = gsub("\nNanoScan", "", pollutant),
    ` ` = paste(rep(" ", 20), collapse = " ")
  )
    
forest(p1_t[,c(1,5:7)],
        est = p1_t$hr,
        lower = p1_t$lower95, 
        upper = p1_t$upper95,
        #sizes = .$se,
        ci_column = 2,
        ref_line = 1,
        arrow_lab = c("Protective", "Detrimental")
        # xlim = c(0, 4),
        # ticks_at = c(0.5, 1, 2, 3),
        )

 



```


```{r}
plot_hrs2 <- function(dt) {
  p <- dt %>%
    ggplot(., aes(y=model, col=pollutant, #col=model,  
                  x=hr, xmin=lower95, xmax=upper95,)) + 
    facet_grid(cols = vars(pollutant),
               #cols = vars(outcome), 
               space="free", scales="free_y", switch = "both") + 
    geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
    geom_pointrange(position = position_dodge(width = 1)) +
    #geom_text(aes(label=hr_label, x=Inf, hjust=1), position = position_dodge(width = 1), size=3)+
    theme(legend.position = "none") +
    labs(x= "HR (95% CI)", y="", col = "Exposure Model", shape="")
  
  return(p)
}

#presentation
# dt = main_hrs
plot_hr3 <- function(dt) {
  p <- dt %>%
    ggplot(., aes(y=model, col=pollutant, #shape = significant, 
                  x=hr, xmin=lower95, xmax=upper95,)) + 
    facet_grid(rows = vars(pollutant),
               #cols = vars(#outcome), 
               #space="free", scales="free_y", 
               switch = "both"
               ) + 
    geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
    geom_pointrange(position = position_dodge(width = 1)) +
    theme(legend.position = "none") +
    #geom_text(aes(label=hr_label, x=Inf, hjust=1), position = position_dodge(width = 1), size=3)+
    labs(x= "HR (95% CI)", y="")
  
  #p
  
  return(p)
}
```



```{r}
main_hrs <- hrs %>%
  filter(grepl("Main", description),
        !grepl("PM2.5|DiSCmini|P-TRAK", pollutant),
        # dont include SP or ST NO2
        grepl("2019 MM", exposure_model)) %>%
  mutate(
    pollutant = gsub("\nNanoScan", "", pollutant),
    pollutant = factor(pollutant, levels = c(paste0("PNC (", pnc_units ," pt/cm3)"), levels(hrs$pollutant))),
    description = ifelse(grepl("Two Pollutant", model), "Primary", "Sensitivity"))

main_hrs %>%
  ggplot(., aes(y=model, x=hr, xmin=lower95, xmax=upper95,)) +
  facet_grid(rows = vars(pollutant), switch = "both") +
  geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
  geom_pointrange(position = position_dodge(width = 1), aes(linetype=description,
                                                            col=description,
                                                            shape=description
                                                            )) +
  geom_text(aes(label=hr_label, x=.65, hjust=.2, fontface="bold"))+
  labs(x= "HR (95% CI)", y="Exposure Model",
       linetype="Analysis", col="Analysis", shape="Analysis"
       )

ggsave(file.path(image_path, "hrs.png"), width = 8, height = 5)




main_hrs %>%
  filter(grepl("Two Pollutant", model)) %>%
  #plot_hrs2()
  ggplot(., aes(y=model, col=pollutant, #col=model,  
                  x=hr, xmin=lower95, xmax=upper95,)) +
  facet_grid(#rows = vars(model),
             cols = vars(pollutant), #switch = "both"
               ) +
  geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
  geom_pointrange(position = position_dodge(width = 1)) +
  theme(legend.position = "none") +
  scale_x_continuous(breaks = seq(0.8, 1.2, by=.1)
    #breaks = c(.8, 1, 1.2)
                     ) +
  
  labs(x= "HR (95% CI)", y="", col = "Exposure Model", shape="")
 
ggsave(file.path(image_path, "presentation", "main_hrs.png"), width = 6, height = 2)

# plots 
plot_hrs2(main_hrs) + facet_grid(~pollutant)

ggsave(file.path(image_path, "presentation", "multi_pollutant_hrs.png"), width = 7, height = 2)


```

### HR table for MS

```{r}
main_hrs %>%
  select(pollutant, Model=model, hr_label) %>%
  pivot_wider(names_from = "pollutant", values_from = "hr_label") %>%
  select(Model, contains("PNC"), everything()) %>%
  kable(caption = "Hazard ratios (95% CI) for the incidence of all-cause dementia by ten-year average pollutant exposure") %>%
  kable_styling()

```

## Other UFP instruments

```{r}
 
ufp_pollutant_levels <- c("10-420 nm, Primary", "20-1,0000 nm", "Onroad 20-1,0000 nm", "10-700 nm",

                      #NS bins
                      "10-100 nm",
                      "10-13 nm",
                      "13-18 nm",
                      "18-24 nm" ,
                      "24-32 nm",
                      "32-42 nm",
                      "42-56 nm",
                      "56-75 nm",
                      "75-100 nm",
                      "100-133 nm",
                      #"133-178 nm",

                      #P-trak differences
                      "20-36 nm",
                      #"36-1k nm",
                      "36-1,000 nm",

                      # #other total UFP instruments
                      # "PNC (1000 pt/cm3)\nP-TRAK", "PNC (1000 pt/cm3)\nDiSCmini",
                      # 
                      # #onroad PTRAK
                      # "PNC (1000 pt/cm3)\nOnroad",

                      #size
                      #"Median Pt Size",
                      "Mean Particle nm"
                      )



ufp_hrs <- hrs %>%
  filter(grepl("Main|PNC|Particle", description),
        grepl("NanoScan|DiSCmini|P-TRAK|PNC|Size", pollutant),
        grepl("Two Pollutant", model),
        
        #don't plot since upper limit is high (~16) - b/c little variability?
        !grepl("NS_154.0", pollutant_predictors)
        ) %>%  
  mutate(
  pollutant = gsub("PNC (1000 pt/cm3)\n", "", pollutant, fixed = T) %>%
   gsub(" (pt/cm3)", "", ., fixed = T) %>%
  gsub("Onroad", "P-TRAK Onroad 20-1,0000 nm", .) %>%
    gsub("36-1k", "36-1,000", .), 
  
  pollutant = ifelse(pollutant == "NanoScan", "NanoScan (10-420 nm), Primary",
                            ifelse(pollutant == "P-TRAK", "P-TRAK (20-1,0000 nm)",
                                   ifelse(pollutant == "Screened P-TRAK", "Screened P-TRAK (36-1,0000 nm)",
                                          ifelse(pollutant == "DiSCmini", "DiSCmini (10-700 nm)", pollutant# NA
                                                 )))),
  pollutant = ifelse(grepl("PNC", pollutant) & grepl("20-36|36-1k", pollutant), gsub("PNC", "P-TRAK", pollutant),
                            gsub("PNC", "NanoScan", pollutant)),
  pollutant = gsub("NanoScan |NanoScan\n|P-TRAK |P-TRAK\n|DiSCmini|DiSCmini |\\(|\\)", "", pollutant),
  pollutant = gsub("Onroad 20-1,0000 nm", "Onroad 20-1,0000 nm", pollutant),
  
  pollutant = factor(ufp_pollutant_levels, levels=ufp_pollutant_levels),
  
  total_ufp = ifelse(grepl("10-700|10-420|20-1,0000|10-100", pollutant), "Total PNC",
                     ifelse(grepl("Mean Particle", pollutant), "Size", "Size-specific PNC")
                     ),
  instrument = ifelse(grepl("10-700|Particle", pollutant), "DiSCmini",
                      ifelse(grepl("20-|36-", pollutant), "P-TRAK", "NanoScan"
                      )),
  instrument = factor(instrument, levels = c("NanoScan", "P-TRAK", "DiSCmini"))
  
         ) #%>% select(instrument, pollutant, total_ufp) %>% View()

# unique(ufp_hrs$pollutant)


# plots 
#print("plot doesn't include 154 bin b/c upper limit is high (~16)")

ggplot(ufp_hrs, aes(y=pollutant, x=hr, xmin=lower95, xmax=upper95, )) +
  #facet_grid(total_ufp~outcome, space="free", scales="free_y", switch = "both") + 
  facet_grid(rows = vars(total_ufp), space="free", scales="free_y", switch = "both") + 
  geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
  geom_pointrange(position = position_dodge(width = 1), aes(col=instrument, #linetype=instrument, 
                                                            shape=instrument
                                                            )) +
  geom_text(aes(label=hr_label, x=-0.3, hjust=.2, fontface="bold"))+
  #theme(panel.grid.major.y = element_blank()) +
  scale_x_continuous(breaks = c(seq(.5, 1.75, by=.25))) +
  labs(x= "HR (95% CI)", y="UFP Model Sensitivity & Secondary Analyes", linetype="Instrument", col="Instrument", shape="Instrument")

ggsave(file.path(image_path, "ufp_analyses.png"), width = 8, height = 8)


#table
ufp_hrs %>%
  #select(pollutant, Model=model, hr_label) %>%
  select('UFP Model' = pollutant, 'HR (95% CI)' = hr_label) %>%
  kable(caption = "Estimated hazard ratio (95% CI) for the incidence of all-cause dementia associated with an increase in ten-year average UFP (__k pt/cm3) exposure level, as determined by exposure prediction models that use different UFP instruments and particle size ranges.") %>%
  kable_styling()


# for presentation
ufp_hrs %>%
  filter(
    !(grepl("NanoScan", pollutant) & !grepl("Primary", pollutant)),
    !grepl("Pt Size", pollutant),
    !grepl("Onroad", pollutant),
    ) %>% 
  
  mutate(
    pollutant = case_when(
      grepl("10-420 nm", pollutant) ~ "NanoScan, 10nm+ (Primary)",
      grepl("10-700", pollutant) ~ "DiSCmini, 10+ nm",
      grepl("20-1,0000", pollutant) & !grepl("Onroad", pollutant) ~ "P-TRAK, 20+ nm",
      grepl("20-36 nm", pollutant) ~ "P-TRAK, 20-36 nm",
      grepl("36-1k nm", pollutant) ~ "P-TRAK, 36+ nm",
      grepl("Onroad 20-1,0000 nm", pollutant) ~ "P-TRAK, 20+ nm, onroad",
      TRUE~pollutant
      ),
    pollutant =  factor(pollutant, levels=c("NanoScan, 10nm+ (Primary)",
                                            "DiSCmini, 10+ nm",
                                            "P-TRAK, 20+ nm",
                                            "P-TRAK, 20+ nm, onroad",
                                            "P-TRAK, 20-36 nm",
                                            "P-TRAK, 36+ nm"
                                            ))
    ) %>%
  
  
  ggplot(., aes(y=pollutant, #col=pollutant, 
                x=hr, xmin=lower95, xmax=upper95,)) + 
  geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
  geom_pointrange(position = position_dodge(width = 1)) +
  theme(legend.position = "none") +
  labs(x= "HR (95% CI)", y="UFP Measure", col = "Exposure Model", shape="")

#ggsave(file.path(image_path, "presentation", "ufp_measures.png"), width = 5, height = 2.5)

```

## Sensitivity Analyses

```{r}
sensitivity_lvs <- c("Main Analysis", 
                     "Restricted Cohort (2010+)",
                     #"Restricted Cohort (Pre 2019)",
                     "Non-Restricted Cohort (1994-2020)",
                     "Baseline Age Adjustment",
                     "Calendar Time Axis & 2yr Age Adjustment", "Calendar Time Axis\n& 2yr Age Adjustment",
                     
                     "No IPW",
                     "Spatiotemporal 2019", "Spatiotemporal"
                     )

sensitivity_hrs <- hrs %>%
  filter((grepl("Restricted|No IPW|Baseline Age Adjustment|Time Axis|Spatiotemporal", description) | 
            (grepl("Main", description) & grepl("Two Pollutant", model) & !grepl("DiSCmini|P-TRAK", pollutant))),
         !grepl("PM2.5", pollutant),
         # drop NO2 sensitivity analyses
         #!(grepl("NO2", pollutant) & exposure_model %in% c("2019 ST", "ST + 2019 ST")),
         #don't need to present this?
         #!grepl("(Pre 2019)", description)
         ) %>% 
  mutate(pollutant = gsub("\nNanoScan", "", pollutant),
         pollutant = factor(pollutant, levels = c(paste0("PNC (", pnc_units ," pt/cm3)"), levels(hrs$pollutant))),
         #pollutant = droplevels(pollutant),
         description = gsub("Calendar Time Axis & 2yr Age Adjustment", "Calendar Time Axis\n& 2yr Age Adjustment", description),
         #description = droplevels(description),
         description = factor(description, levels = sensitivity_lvs)
         ) 

# plots 
ggplot(sensitivity_hrs, aes(y=description, #col=pollutant, 
                            x=hr, xmin=lower95, xmax=upper95,)) + 
    facet_grid(rows=vars(pollutant), #cols = vars(outcome), 
               space="free", scales="free_y", switch = "both"
               ) + 
    geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
    geom_pointrange(position = position_dodge(width = 1)) +
    geom_text(aes(label=hr_label, hjust=0.08,  x=.7, fontface="bold"))+
  theme(legend.position = "none") +
    labs(x= "HR (95% CI)", y="Sensitivity Analyses", #col = "Exposure Model", 
         shape="")

ggsave(file.path(image_path, "SI", "sensitivity_analyses.png"), width = 8, height = 8)


sensitivity_hrs %>%
  select(description, pollutant,  hr_label) %>%
  pivot_wider(names_from = "pollutant", values_from = "hr_label") %>%
  select(description, contains(c("PNC", "BC", "NO2"))) %>%
  #mutate(description = factor(description, levels = sensitivity_lvs)) %>%
  arrange(description) %>%
  rename('Other Sensitivity Analysis'= description) %>%
  kable(caption = "Other sensitivity analyses of the estimated hazard ratio (95% CI) for the incidence of all-cause dementia associated with an increase in ten-year average pollutant exposure level.") %>%
  kable_styling() %>% 
  add_footnote("*: p<0.05")

```

 

## Secondary Analyses

```{r}
secondary_hrs <- hrs %>%
  filter(grepl("Subtype Outcomes", description) |
            grepl("Shorter Exposure", description) |
           (grepl("Main", description) & !grepl("DiSCmini|P-TRAK|PM2.5", pollutant) & grepl("Two Pollutant", model)),
         !grepl("PM2.5", pollutant),
         # drop NO2 sensitivity analyses
         !(grepl("NO2", pollutant) & exposure_model %in% c("2019 ST", "ST + 2019 ST"))
         ) %>% 
  mutate(pollutant = gsub("\nNanoScan", "", pollutant),
          pollutant = factor(pollutant, levels = c(paste0("PNC (", pnc_units ," pt/cm3)"), levels(hrs$pollutant))),
         description = ifelse(grepl("Subtype", description), paste(outcome, "Outcome"),
                              ifelse(grepl("Exposure Period", description), paste(exposure_duration, "Yr Exposure Period"),
                                     as.character(description))),
         description = relevel(factor(description), ref = "Main Analysis"),
         exposure_duration = factor(exposure_duration, levels = c(1,5,10))
         ) 
#secondary_hrs %>% arrange(pollutant) %>% select(pollutant, outcome, exposure_duration, description)

# plot [same as above]
ggplot(secondary_hrs, aes(y=description, #col=pollutant, 
                          x=hr, xmin=lower95, xmax=upper95)) + 
  facet_grid(rows=vars(pollutant), space="free", scales="free_y", switch = "both") + 
  geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
  geom_pointrange(position = position_dodge(width = 1)) +
  #theme(legend.position = "none") +
  geom_text(aes(label=hr_label, hjust=0.08,  x=.75, fontface="bold"))+
  labs(x= "HR (95% CI)", y= "Secondary Analyses", 
       col = "Outcome", shape="")

ggsave(file.path(image_path, "SI", "secondary_analyses.png"), width = 8, height = 8)
#ggsave(file.path(image_path, "presentation", "secondary_analyses.png"), width = 6, height = 3)


# table
secondary_hrs %>%
  select(description, pollutant,  hr_label) %>%
  pivot_wider(names_from = "pollutant", values_from = "hr_label") %>%
  select(description, contains(c("PNC", "BC", "NO2"))) %>%
    rename('Secondary Analysis'= description) %>%

  kable(caption = "Secondary analyses of the estimated hazard ratio (95% CI) for the incidence of dementia subtype (AD or non-AD) associated with an increase in 1- or 5-year average pollutant exposure level.The main analysis is included for reference. It models all-cause dementia incidence using a 10-year average pollutant exposure level") %>%
  kable_styling()  


```


# Exploratory - interaction models


```{r, eval=F}
# --> add units 

modification_hrs <- hrs_em %>%
  mutate(
    em= ifelse(grepl("not_var", covariate), "no", "yes"),
         # description = gsub(" Interaction", "", description),
         # description = gsub("_", " ", description),
         # description = str_to_title(description),
         # description = gsub("Race ", "Race, ", description),
         # description = ifelse(grepl("apoe", description, T), "APOEe4", description),
         covariate = gsub("_var|_not_var|MM_", "", covariate),
    
         group_variable = ifelse(grepl("male", description) & em=="yes", "Male",
                               ifelse(grepl("male", description) & em=="no", "Female",
                                      ifelse(grepl("race", description) & em=="yes", "White",
                                             ifelse(grepl("race", description) & em=="no", "People of Color",
                                                    ifelse(grepl("apoe", description) & em=="yes", "Carrier",
                                                           ifelse(grepl("apoe", description) & em=="no", "Non-Carrier",
                                                                  ifelse(grepl("degree", description) & em=="yes", "Higher Degree",
                                                                         ifelse(grepl("degree", description) & em=="no", "Lower Degree",
                                                                                ifelse(grepl("low_ndi", description) & em=="yes", "Lower NDI",
                                                                                       ifelse(grepl("low_ndi", description) & em=="no", "Higher NDI", NA
                                                                                       )))))))))
                               ),
    
    
    group_description = ifelse(grepl("male", description, ignore.case = T), "Sex",
                               ifelse(grepl("race", description), "Race",
                                      ifelse(grepl("apoe", description), "APOEe4",
                                             ifelse(grepl("degree", description), "Educational Degree",
                                                    ifelse(grepl("ndi", description), "Neighborhood Deprivation Index", NA))))),
    ) %>%
  select(group_description, group_variable, pollutant=covariate, hr_label, hr, lower95, upper95) %>%
label_pollutants() %>%
  mutate(pollutant = gsub(", NanoScan", "", pollutant),
         #pollutant = factor(pollutant, levels = c("PNC (1000 pt/cm3)", levels(hrs$pollutant)))
         ) 


#plot
modification_hrs %>%
  filter(grepl("race|sex", group_description, ignore.case = T)) %>%
  
  ggplot(., aes(y=group_variable, col=pollutant, x=hr, xmin=lower95, xmax=upper95)) + 
    facet_grid(rows=vars(group_description), cols = vars(pollutant), 
               space="free", scales="free_y", #switch = "both"
               ) + 
    geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
    geom_pointrange(position = position_dodge(width = 1)) +
  theme(legend.position = "none")+
    labs(x= "HR (95% CI)", y="Effect Modification", col = "Pollutant", shape="")

ggsave(file.path(image_path, "presentation", "effect_modification1.png"), width = 5.5, height = 3)

 modification_hrs %>%
  filter(!grepl("race|sex", group_description, ignore.case = T)) %>%
  
  ggplot(., aes(y=group_variable, col=pollutant, x=hr, xmin=lower95, xmax=upper95)) + 
    facet_grid(rows=vars(group_description), cols = vars(pollutant), 
               space="free", scales="free_y", #switch = "both"
               ) + 
    geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
    geom_pointrange(position = position_dodge(width = 1)) +
  theme(legend.position = "none")+
    labs(x= "HR (95% CI)", y="Effect Modification", col = "Pollutant", shape="")

ggsave(file.path(image_path, "presentation", "effect_modification2.png"), width = 5.5, height = 7)




# table
modification_hrs %>%
  select(-c("hr", "lower95", "upper95")) %>%
  pivot_wider(names_from = c("pollutant"), values_from = "hr_label") %>%
  rename('Effect Modifier' = group_description,
         ' ' = group_variable) %>%
  kable(caption = "Interaction models") %>%
  kable_styling()
```



# PM2.5 Results with new data & models - using SP 2019 & ST PM2.5  

* ST PM2.5 in single pollutant model here is similar but not identical to Rachels original fidings. This is a different analysis

* other PM2.5 results are for SP 2019 PM2.5, so not what we would report in a paper 

* note that the PM2.5 effect is for 3 instruments (NS, P-TRAK, discmini)

```{r}
keep_main_st_pm25 <- hrs %>%
  filter(pollutant_predictors == "PNC + PM2.5",
         description == main_models
         ) %>%  
  # the PM2.5 estimate we want is the one that follows the NS estimate
  slice(grep("NanoScan", pollutant)+1)

hrs %>%
  filter(grepl("Main", description),
        grepl("+PM2.5", model),
        exposure_duration == 10,
        # dno't want SP or ST NO2
        !(pollutant_predictors == "NO2 + PM2.5" & exposure_model =="2019 ST"),
        !(pollutant_predictors == "NO2 + PM2.5" & exposure_model == "ST + 2019 ST"),
        !grepl("DiSCmini|P-TRAK", pollutant),
        #will only use 1 SP PM2.5 estimate, from the NS instrument model
        !(pollutant_predictors == "PNC + PM2.5" & grepl("PM2.5", pollutant))
        ) %>% 
  bind_rows(keep_main_st_pm25) %>%
  mutate(
    description = "SP PM2.5",
    analysis = "Main"
  ) %>%
  bind_rows(hrs_st_pm25 %>% mutate(analysis = "Sensitivity")) %>%  
  #don't include these models?
   filter(pollutant_predictors != "PNC + BC + NO2 + PM2.5") %>%
  mutate(
    pollutant_predictors = factor(pollutant_predictors, levels = unique(hrs$pollutant_predictors)),
    pollutant = factor(pollutant, levels = c(pollutant_levels))
    ) %>%
 
  ggplot(., aes(y=pollutant, x=hr, xmin=lower95, xmax=upper95, col=analysis)) +
  facet_grid(pollutant_predictors~description, scales = "free", switch="y") +
  
  geom_vline(xintercept = 1, alpha=0.5, linetype=2) +
  geom_pointrange(position = position_dodge(width = 1), aes( )) +
  geom_text(aes(label=hr_label, x=.2, hjust=0, fontface="bold"))+
  labs(x= "HR (95% CI)", y= "Sensitivity Analyses with PM2.5",
       col = "Analysis"
       )

ggsave(file.path(image_path, "SI", "hrs_sp_st_pm2.5.png"), width = 7, height = 6)


```



# ST PM2.5 results using the NEW DATASET WITHOUT THE MM COVERAGE REQUIREMENT (WHAT WE MIGHT PUBLISH AFTER ADDING AN ST PM2.5 COVERAGE VARAIBLE)
* results we would get if we re-did Rachel's analysis 

* uses model adjustment similar to current models (e.g., NDI, IPW)

* results are similar to Rachel's work

```{r}
st_pm2.5_hrs0 <- readRDS(file.path(output_data_path, "hazard_ratios_st_pm2.5.rda")) 

st_pm2.5_hrs <- st_pm2.5_hrs0 %>%
  filter(covariate == "ST_pm25") %>%
  mutate(
    significant = ifelse((lower95 <1 & upper95<1) | (lower95 >1 & upper95>1), "Significant", "Non-Significant" ),
    hr_label = paste0(format(round(hr, 2), nsmall=2), " (", format(round(lower95, 2), nsmall=2), "-", format(round(upper95,2), nsmall=2), ")"),
    )

print("PM2.5 effects, without adjustment, or with adjustment from MM UPF, NO2 and/or BC models")
st_pm2.5_hrs %>%
  ggplot(., aes(y=pollutant_predictors, x=hr, xmin=lower95, xmax=upper95,)) +
  facet_grid(~ "All-Cause Dementia", #vars(outcome),
             #cols = vars(exposure_model),
             switch = "both") +
  geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
  geom_pointrange(position = position_dodge(width = 1), aes(#linetype=description,
                                                            # col=description,
                                                            # shape=description
                                                            )) +
  geom_text(aes(label=hr_label, x=.85, hjust=.2, fontface="bold"))+
  labs(x= "HR (95% CI)", y="ST PM2.5 Model",
       linetype="Analysis", col="Analysis", shape="Analysis",
       title = "Dementia incidence HRs for ST PM2.5, adjusted for MM exposures"
       )

ggsave(file.path(image_path, "Other", "PM2.5", "hrs_st_pm2.5.png"), width = 8, height = 5)

```

same as above but showing effect estiates for other pollutants.

* results we would have seen if we had used ST PM.5 in the models     
* why does BC look protective? PM2.5 & BC are somewhat correlated. Is time-varying PM2.5 soaking up the effect of BC since BC has high expsure measurement error for early years?

```{r}
st_pm2.5_hrs0 %>%
  filter( grepl("ST_pm25|no2|bc|ufp", covariate)) %>% 
  mutate(
    significant = ifelse((lower95 <1 & upper95<1) | (lower95 >1 & upper95>1), "Significant", "Non-Significant" ),
    hr_label = paste0(format(round(hr, 2), nsmall=2), " (", format(round(lower95, 2), nsmall=2), "-", format(round(upper95,2), nsmall=2), ")"),
    ) %>%

#st_pm2.5_hrs %>%
  ggplot(., aes(y=covariate, x=hr, xmin=lower95, xmax=upper95,)) +
  facet_wrap(#~ "All-Cause Dementia", #vars(outcome),
    ~pollutant_predictors, scales="free"
             # rows = vars(covariate),
             # cols = vars(pollutant_predictors),
             # switch = "both"
    ) +
  geom_vline(xintercept = 1, alpha=0.5, linetype=2) + 
  geom_pointrange(position = position_dodge(width = 1), aes(#linetype=description,
                                                            # col=description,
                                                            # shape=description
                                                            )) +
  geom_text(aes(label=hr_label, x=.6, hjust=.2, fontface="bold"))+
  labs(x= "HR (95% CI)", y="ST PM2.5 Model",
       linetype="Analysis", col="Analysis", shape="Analysis"
       )

ggsave(file.path(image_path, "Other", "PM2.5", "hrs_st_pm2.5_v2.png"), width = 18, height = 5)

```
